Moving http_trigger ops to /trigger_http package
